requirements.txt
# 섪치한 라이브러리를 작성해주세요.
# ex) numpy==1.24.3
# ex) pandas==2.0.3
# ex) scikit-learn==1.3.0
torch
numpy
scikit-learn
websocket-client